+++ windows/flutter/generated_plugins.cmake
@@ -6,7 +6,6 @@ list(APPEND FLUTTER_PLUGIN_LIST
   app_links
   audioplayers_windows
   flutter_inappwebview_windows
-  printing
   share_plus
   sqlite3_flutter_libs
   url_launcher_windows